refactorisation des controller pour l'interface bar
